Added Mux
This was not easy and required quite a bit of thinking. I also needed
to search up the actual solution. I was very close to the answer but
was thinking about the bits in from a high level perspective when
really I should have just been focusing on their individual values and
transforming said values into logic.